requirements.txt
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy==1.8.1
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.2.0
django-formtools==2.2
django-guardian==2.2.0
django-import-export==2.0.2
django-js-asset==1.2.2
django-reversion==3.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.17.3
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.2.1
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
zope.interface==5.1.0
